refactor(proto): move proto go files alongside proto files
This commit also bans importing the api/go package from pkg/
