+++ api/go/executor.go
@@ -20,7 +20,7 @@ import (
 	goplugin "github.com/hashicorp/go-plugin"
 	slogctx "github.com/veqryn/slog-context"
 
-	bonkv0 "go.bonk.build/api/go/proto/bonk/v0"
+	bonkv0 "go.bonk.build/api/proto/bonk/v0"
 	"go.bonk.build/pkg/executor"
 	"go.bonk.build/pkg/task"
 )
