+++ api/go/log_streaming.go
@@ -24,7 +24,7 @@ import (
 	slogmulti "github.com/samber/slog-multi"
 	slogctx "github.com/veqryn/slog-context"
 
-	bonkv0 "go.bonk.build/api/go/proto/bonk/v0"
+	bonkv0 "go.bonk.build/api/proto/bonk/v0"
 )
 
 var (
