+++ pkg/executor/rpc.go
@@ -11,7 +11,7 @@ import (
 
 	"google.golang.org/protobuf/types/known/structpb"
 
-	bonkv0 "go.bonk.build/api/go/proto/bonk/v0"
+	bonkv0 "go.bonk.build/api/proto/bonk/v0"
 	"go.bonk.build/pkg/task"
 )
